enhancing boot notification service tests: reverting pulling out of set variables since it's not as easy to do as initially thought
