+++ 03_Modules/Configuration/src/module/DeviceModelService.ts
@@ -144,63 +144,4 @@ export class DeviceModelService {
 
     await Promise.all(promises);
   }
-
-  async executeSetVariablesAfterBoot(
-    stationId: string,
-    cacheCallback: (correlationId: string) => Promise<string | null>,
-    sendSetVariableRequest: (
-      correlationId: string,
-      data: SetVariableDataType[],
-      itemsPerMessageSetVariables: number,
-    ) => Promise<IMessageConfirmation>,
-  ): Promise<[boolean, boolean]> {
-    let rejectedSetVariable = false;
-    let rebootSetVariable = false;
-    let setVariableData: SetVariableDataType[] =
-      await this._deviceModelRepository.readAllSetVariableByStationId(
-        stationId,
-      );
-
-    // If ItemsPerMessageSetVariables not set, send all variables at once
-    const itemsPerMessageSetVariables =
-      (await this.getItemsPerMessageSetVariablesByStationId(stationId)) ??
-      setVariableData.length;
-
-    while (setVariableData.length > 0) {
-      const correlationId = uuidv4();
-
-      await sendSetVariableRequest(
-        correlationId,
-        setVariableData,
-        itemsPerMessageSetVariables,
-      );
-
-      setVariableData = setVariableData.slice(itemsPerMessageSetVariables);
-
-      const setVariablesResponseJsonString = await cacheCallback(correlationId);
-
-      if (setVariablesResponseJsonString) {
-        if (rejectedSetVariable && rebootSetVariable) {
-          continue;
-        }
-
-        const setVariablesResponse: SetVariablesResponse = JSON.parse(
-          setVariablesResponseJsonString,
-        );
-        setVariablesResponse.setVariableResult.forEach((result) => {
-          if (result.attributeStatus === SetVariableStatusEnumType.Rejected) {
-            rejectedSetVariable = true;
-          } else if (
-            result.attributeStatus === SetVariableStatusEnumType.RebootRequired
-          ) {
-            rebootSetVariable = true;
-          }
-        });
-      } else {
-        throw new Error('SetVariables response not found');
-      }
-    }
-
-    return [rejectedSetVariable, rebootSetVariable];
-  }
 }
